requirements.txt
anyio==4.8.0
certifi==2025.1.31
h11==0.14.0
h2==4.2.0
hpack==4.1.0
httpcore==1.0.7
httpx==0.28.1
hyperframe==6.1.0
idna==3.10
pip==24.0
python-telegram-bot==22.1
pytz==2025.2
PyYAML==6.0.2
sniffio==1.3.1
socksio==1.0.0
telegram==0.0.1
typing_extensions==4.12.2
